Debug the zeros_like, zeros, dtype_id, rename the getShapeTensor.
1. zeros_like.
The method zeors_like can't init a tensor with the same type like the given ts.

2. zeros.
To fix the '1', the zeros has been fixed with a new template version. Also, the constructor of tensor also extends a new template with given typename T.

3. getShapeTensor
The name here I think is a little bit ambiguous. So I rename it by 'init_with_shape', which means initialize a tensor with given shape.

4. Autograd framework.
Just leant the framework of Autograd. The framework has been tested on 'double' value, still need to tested at the 'Tensor' object.